Unset cscopetag as it interferes with CtrlPTag
After quite a bit of debugging, and with some help of the author of
CtrlP, we were able to track down why CtrlPTag was popping up a list all
the time after selecting which tag to jump to.  It turned out that
cscopetag was intefering with the normal operation, and would cause the
list to pop up even though CtrlP was trying to avoid triggering the
list.  This fixes the issue for now, but hopefully a better fix will
come along later.