Make CtrlPTag center the tag in the window.
At the moment, it leaves the cursor on the tag, but it could be anywhere
in the window.  This patch puts the tag right in the center, where you
expect it.  I'm trying to this upstream, but the maintainer doesn't seem
to think it's useful when the tags file is out-of-date.  He had this to
say:

  If :tag fails, running zvzz blindly will open the folds and change
  the view of the current buffer.

I disagree though.  First, Vim searches to try and locate the tag.  I
even experimented and examined the source to make sure that it did so.
Second, if :tag fails it still moves the cursor, so it's still helpful
to see where you are.  Finally, having the cursor appear in a known
location on the screen, with an equal chance of the tag being above or
below, is a Good Thing even if the tags are off a bit.