Allow users to disable the use of Powerline.
While I think it's a great default, it may not be for everyone.  So
provide a way for users to opt-out of having it, if they prefer the
default status line.

I updated my configuration as well, since I like being able to debug
things without Powerline in the mix.