Fix method of disabling Powerline plugin.
Vim's convention is that g:PluginName_loaded is a variable used by a
plugin to provide one-time initialization at startup, typically using
this idiom in the plugin's main file:

  if exists("g:PluginName_loaded")
      finish
  endif
  let g:PluginName_loaded = 1

Because plugins are initialized after the main vimrc has run, it's
possible to disable an installed plugin by setting this variable in
advance.

It's tempting to use this variable in the vimrc to indicate whether the
plugin is going to be active, something like this:

  if !exists("g:PluginName_loaded")
      " Plugin has not been disabled, so it will be active.
      ....
  endif

The problem is that this check works only the first time the vimrc is
sourced.  It's desirable that the user be able to make minor changes to
the vimrc and :source it again to have the changes take effect
immediately without restarting Vim.  Some changes will require a restart
anyway (such as enabling or disabling a complex plugin like Powerline),
but as much as possible it's desirable to allow vimrc to be re-sourced.

For this particular case, we want to allow the user to disable
Powerline.  If Powerline is disabled, we further want to avoid that
toggling the Tagbar window causes an update of the Powerline status
line.  If the above technique were used, the problem would be that the
Tagbar plugin's behavior would change on the second :source of vimrc,
because the variable g:Powerline_loaded goes from not existing to being
defined as 1.  For this reason, a separate control variable is
introduced (g:EnablePowerline) that will remain unchanged after the
first sourcing of vimrc.  This variable is then used to determine
whether toggling the Tagbar window should also update the Powerline
status.