Add some aliases for the headings, and stopped jumping out of blocks.
Jumping out of the code block was particularly annoying as you would
inevitable hit tab while typing some hunk of code.  So stop doing that,
and fix up the others for consistency.  Also, add a line between the
directive and where the text goes.  This is the more canonical form, and
I've run into instances where not having it doesn't work (or hit a bug).