Set iskeyword while editing help files.
It got frustrating that it wasn't easy to test links to other
documentation because iskeyword was not set the same way as when you are
viewing help files via :help.  This sets iskeyword so that it matches
how it's set while viewing help so that jumping to links like
'textobj-indent.txt' work correctly.