Disable the error highlighting.
It turns out the error highlighting is not very useful.  It's fixed at a
specific line and column.  If you make changes elsewhere, the
highlighting seemingly moves with the text, but it actually doesn't.  Do
to a redrawing bug in Vim, you'll eventually see several bits
highlighted.  Only when you do a CTRL-L to refresh, do you see what's
really supposed to be highlighted, and it's wrong because the text has
changed position.  Just turn it off.