Fix opening the command-line window programmatically.
It turns out that even though <C-G> and <C-O><C-F> where mapped to
<C-F>, cedit was still set to <C-F>--which was mapped to <Right>.
Therefore, any plugin that was trying to use the cedit sequence to open
the command line window, was failing.  But, it did work when the user
typed it at the command line.  Trying to change cedit to be <C-G> or
<C-O><C-F> didn't work either.  Instead of popping up the command line
window, it would simply put in a literal CTRL-F instead, even when
invoked via a plugin.

So, let's remove the <C-G> mapping, make <C-O><C-F> insert a <C-G>, and
change cedit to be <C-G>.  This not only allows our customization to
continue working, but then cedit is set correctly and plugins that use
it can pop open the command line window (such as vim-fireplace).