sniputils: make the flags to b/wsnip and b/wabbr more useful
Setting the default value to include the "b" or "w" means that you must
specify the "b" or "w' yourself when trying to pass additional flags,
such as "!".  If you have to specify them both, then there doesn't seem
to be much value over using snip() and abbr() directly.  Instead, let
the default flags value be empty, and add in the "b" or "w" flag, so
that the use can just pass in the additional flags that they want.
Also, run a replace operation to remove the flag from the passed in
value, just to prevent any issues.