Remove VIMRC_BUNDLE concept.
VIMRC_BUNDLE was intended to provide an area for per-use bundles, but
pathogen#infect(path) has an unexpected behavior for paths that contain
slashes.  Such paths will surround (via pathogen#surround()) the
runtimepath, causing them to come earlier in runtimepath than the user's
own ~/.vim directory.  As a side-effect, the user's spellfile would be
created in a bundle directory because Vim defaults to putting the
spellfile into the first writable area in runtimepath.

Because the vimrc already puts the user's $VIMUSERFILES area in the
runtimepath, there is no need to have a separate VIMRC_BUNDLE mechanism,
since any bundles living in $VIMUSERFILES/bundle will be handled by the
general pathogen#infect() call; therefore, the VIMRC_BUNDLE concept is
being removed.