Update to a newer version of UltiSnips.
While trying to support templates for new buffers, I discovered that the
UltiSnips_SnippetsInCurrentScope() function was busted when the
description contained an apostrophe.  My fix was incorporated upstream,
and this version contains that fix.  The new version also contains
support for buffer local snippet directories, so this removes the
autocmd workaround we had in place.