Add support for indent guides.
vim-indent-guides wants to work globally by default, using a map to say
which files you don't want guides for.  This setup turns that around and
provides a buffer-local mechanism to turn indent guides on and off.  It
also implements the same scheme as `g:SpellMap` to aid in enabling
indent guides for specific filetypes and groups.  Currently, only Python
is turned on, and only if indent guides is enabled globally by setting
`g:IndentGuides` to 1.