Make the default filetype for Markdown files be "markdown".
It appears "markdown" is a more common filetype for markdown files, so
it seems to be a better default.  To provide some support for mkd, just
in case, mkd snippets now delegates to markdown.

This change also means that anyone with custom mkd snippets and
autocommands needs to change them to markdown instead.

Also, we now use vim-markdown's filetype detection.  Now .mkdn is
recognized as a markdown filetype and now .mdwn is not.  This seems
reasonable since .mdwn seems unlikely to be used.