Avoid formatoptions+=ro for text files.
The 'r' and 'o' flags are documented to do only one thing: automatically
insert the comment leader when pressing <Enter> and o/O, respectively.
But they seem to have a side-effect: when these flags are present,
bulleted lines cause extra indentation on the next line. E.g.:

  - This bulleted will indent to
    here after pressing <Enter>.

For text files, we're leaving out the 'r' and 'o' for now.  Mostly,
there aren't comment characters in such files, though Markdown and
others have defined 'comments' to be certain useful things, and it might
be nice to re-enable these flags someday if the indentation issue can be
sorted out.