Add the ability to "inherit" from a directory.
This it to help with folks who want to derive their configuration from
another person's.  It will insert the inherited directory just after
your own in the runtimepath, and insert the after/ directory in the
correct location, if it exists.  Then you can add the following to your
vimrc-before.vim:

    call InheritDirectory(expand("$VIMFILES/user/name"))
    exec "source " . expand("$VIMFILES/user/name/vimrc-before.vim")

and the following to your vimrc-after.vim:

    exec "source " . expand("$VIMFILES/user/name/vimrc-after.vim")

where "name" is the name of the user's setting you want to inherit.