Add a version search highlighting with more contrast to szakdark.
Default to this new version of search highlighting.  User's can get the
more subtle technique used before by setting `g:szakdark_subtle_search`
to 1 before setting the colorscheme.

This also drops the use of underline in the subtle search, along with
bumping up the background color slightly, to make it standout a little
more, and providing a better console highlighting for search terms--the
prior version was abysmal!