Patch UltiSnips to avoid crashing when restoring the clipboard.
It turns out that you can end up with some interesting content in the
clipboard that cannot be converted to Unicode--in my case there were
some '\x80` bytes and it terminated with an escape sequence.  UltiSnips
catches this and just returns the string instead.  However, in the path
to restore the clipboard, `_vim_enc()` is called with the string, and
attempts to encode it with Vim's current encoding setting.  It turns
out, this has an interesting affect with byte strings.  Python attempts
to convert it to Unicode first, causing a `UnicodeDecodeError` to be
thrown, which is not handled by `_vim_dec()`, so it leaves UltiSnips in
a terrible state.

Let's at least attempt to catch the error, and return he string directly
(as with a `UnicodeEncodeError`) and let UltiSnips complete its work.