Scale the font size under OS X.
This adds an AdjustBaseFontSize() function to help scale the font size
for use under Mac OS X.  The issue is that Mac's idea of a point is not
the same definition as the rest of the world (despite having so much
heritage in the print industry).  As a result, we need to scale it to
match the typical monitor DPI.  This works for Retina displays, since OS
X has another scaling factor that it's applying there too.  You don't
get actual point sizes from this (unless you're at the ideal scaling
factor for a retina display), but it does match expected behavior of
font sizes on OS X.