Fix a powerline bug where symbols would disappear from the dictionary.
While trying to work on a replacement fileinfo segment that does not use
a bolded LN symbol (to make it work better with the Hack typeface), I
noticed that the BRANCH and LINE symbols would eventually become [''].
It's still not entirely clear why this is happening, but it smells like
something is mutating a reference.  As a result, let's create a copy of
the arg in Pl#Parser#ParseChars() and return that instead.  This fix
allows everything to work correctly again.