Set `loaded_airline` to disable airline
In the logic to disable airline, `g:Airline_loaded` was set in an
attempt to prevent the plugin from loading which was not working as
expected.  Looking through the plugin, loading is controlled via
`g:loaded_airline`.  This request updates the logic to set the
appropriate variable to prevent airline from loading.