Make IsQuickFixWin() work correctly in newer versions of Vim.
At some point in Vim 8.0 they stopped allowing the :lopen command from
being called inside an autocmd.  Unfortunately, this causes an error
trace and it doesn't fail in a way that allows us to distinguish the
type of window it is either.  However, a new built-in was added to allow
us to get the window id and look up whether it's a quickfix or loclist.
So we'll prefer that technique when available, and when it isn't we fall
back to the old technique.  win_getid() first appeared in 7.4.2063, so
there should not be any gaps where both techniques may fail.