Try built-in ``gf`` functionality before falling back to ``tags`` file.
The built-in ``gf`` command has a number of fancy ways to find source
files.  Previously, ``gf`` was mapped to use tags files to lookup file
names, but this loses out on some built-in settings for ``gf``.  So
first try out the built-in ``gf`` command, and if that fails to locate a
filename, fall back to the tags-based approach.